test: move gdalinfo test
