+++ packages/topo-imagery-gdal/test/tiff/file_tiff_test.py
@@ -1,5 +1,5 @@
+from gdalinfo import add_band, add_palette_band, fake_gdal_info
 from topo_imagery_gdal.gdal.gdal_presets import CompressionPreset
-from topo_imagery_gdal.testing.gdalinfo import add_band, add_palette_band, fake_gdal_info
 from topo_imagery_gdal.tiff.file_tiff import FileTiff, FileTiffErrorType
